+++ example/example/urls.py
@@ -1,6 +1,7 @@
 from app.views import index
 from app.views import pdf_file
 from app.views import stream
+from app.views import terminal_controls
 from app.views import ugly_json
 from app.views import ugly_xml
 from app.views import unauthorized
@@ -11,6 +12,7 @@ urlpatterns = [
     path("401", unauthorized),
     path("pdf_file", pdf_file, name="pdf_file"),
     path("stream", stream, name="stream"),
+    path("terminal_controls", terminal_controls, name="terminal_controls"),
     path("ugly_json", ugly_json, name="ugly_json"),
     path("ugly_xml", ugly_xml, name="ugly_xml"),
 ]